split AccountViewModel into 2 partial classes
