+++ FinanceTracker/FinanceTracker.WPF/ViewModels/AccountViewModel_Summary.cs
@@ -10,7 +10,8 @@ using FinanceTracker.Core;
 
 namespace FinanceTracker.WPF
 {
-    public class AccountViewModel : ViewModelBase<AccountModel>
+    //summary tab of accountViewModel
+    public partial class AccountViewModel : ViewModelBase<AccountModel>
     {
         public string AccountName { get { return _m.Name; } }
         public string CurrencyType { get { return _m.CurrencyType; } }
@@ -24,18 +25,14 @@ namespace FinanceTracker.WPF
         private double _averagePurchaseCost = 0.0;
         public double AveragePurchaseCost { get { return _averagePurchaseCost; } private set { _averagePurchaseCost = value; Notify(); } }
 
-        private string _rawTransFilePath;
-        public string RawTransFilePath { get { return _rawTransFilePath; } set { _rawTransFilePath = value; Notify(); } }
-
         public ObservableCollection<CategoryStats> Categories { get; set; } = new ObservableCollection<CategoryStats>();
         public ObservableCollection<TransactionModel> Transactions { get; set; } = new ObservableCollection<TransactionModel>();
-        public ObservableCollection<ConvertTransactionViewModel> ConvertTransactions { get; set; } = new ObservableCollection<ConvertTransactionViewModel>();
         public LamdaCommand AddTransactions { get; set; }
         public LamdaCommand UpdateBalance { get; set; }
-        public LamdaCommand BrowseCmd { get; set; }
         public AccountViewModel() : base() { }
         public AccountViewModel(AccountModel model) : base(model)
         {
+            Initialize_Add();
         }
 
         protected override async void Initialize()
@@ -52,11 +49,6 @@ namespace FinanceTracker.WPF
             UpdateBalance = new LamdaCommand(
                 (obj) => true,
                 (obj) => ShowBalanceDialog()
-            );
-
-            BrowseCmd = new LamdaCommand(
-                (obj) => true,
-                (obj) => BrowseForFile()
             );
 
             Notify(nameof(AccountName));
@@ -123,38 +115,5 @@ namespace FinanceTracker.WPF
                 TransactionModel? tm = await SQLiteContext.AddTransaction(trans);
             }
         }
-
-        private async void BrowseForFile()
-        {
-            var dialog = new Microsoft.Win32.OpenFileDialog();
-            bool? result = dialog.ShowDialog();
-
-            if (result == true)
-            {
-                RawTransFilePath = dialog.FileName;
-
-                string fileText = File.ReadAllText(RawTransFilePath);
-                string[] lines = fileText.Split('\n');
-
-                var splitterRule = await SQLiteContext.GetConversionRule_Splitter(_m.Id);
-                var nameRule = await SQLiteContext.GetConversionRule_Name(_m.Id);
-                var dateRule = await SQLiteContext.GetConversionRule_Date(_m.Id);
-                var dollarValueRule = await SQLiteContext.GetConversionRule_DollarValue(_m.Id);
-                var categoryRule = await SQLiteContext.GetConversionRule_Category(_m.Id);
-                var balanceRule = await SQLiteContext.GetConversionRule_Balance(_m.Id);
-
-                ConvertTransactions.Clear();
-                foreach (string line in lines)
-                {
-                    if (line == "")
-                        continue;
-
-                    ConvertTransactionViewModel ct = new ConvertTransactionViewModel(_m, line);
-                    await ct.ConvertTransaction(splitterRule, nameRule, dateRule, dollarValueRule, balanceRule, categoryRule);
-                    ConvertTransactions.Add(ct);
-                }
-            }
-        }
-
     }
 }
